Moved non-API views out of the API * Views that make up APIv1 now in api/v1/ * Channels handlers moved to channels app * Stripe handler moved to orgs app * Old api/v1/xxx URLs kept for backwards compatibility
