+++ temba/api/v1/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import absolute_import, unicode_literals
+
+"""
+API v1. This will be deprecated once v2 is complete, and eventually removed
+"""
