+++ temba/urls.py
@@ -22,12 +22,12 @@ urlpatterns = patterns('',
     url(r'^', include('temba.campaigns.urls')),
     url(r'^', include('temba.ivr.urls')),
     url(r'^', include('temba.locations.urls')),
-    url(r'^channels/', include('temba.channels.urls')),
+    url(r'^', include('temba.api.urls')),
+    url(r'^', include('temba.channels.urls')),
     url(r'^relayers/relayer/sync/(\d+)/$', sync, {}, 'sync'),
     url(r'^relayers/relayer/register/$', register, {}, 'register'),
     url(r'^users/', include('smartmin.users.urls')),
     url(r'^imports/', include('smartmin.csv_imports.urls')),
-    url(r'^api/v1', include('temba.api.urls')),
     url(r'^assets/', include('temba.assets.urls')),
     url(r'^jsi18n/$', 'django.views.i18n.javascript_catalog', js_info_dict)
 )
